+++ src/client/common/infinite_board.test.ts
