+++ src/client/common/infinite_board.ts
