+++ src/client/games/slashchain/rules.ts
