+++ src/client/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App";
 import { Route } from "react-router";
-import { BrowserRouter } from "react-router-dom";
+import { BrowserRouter } from "r./client/games/Debugom";
 
 ReactDOM.render(
   <React.StrictMode>
